[issue-456] add helper method in casing_tools and move the module as it is used in the parser and writer layer
