+++ src/spdx/writer/rdf/annotation_writer.py
@@ -14,7 +14,7 @@ from rdflib import Graph, Literal, RDFS, URIRef, RDF, BNode
 
 from spdx.datetime_conversions import datetime_to_iso_string
 from spdx.model.annotation import Annotation
-from spdx.writer.casing_tools import snake_case_to_camel_case
+from spdx.casing_tools import snake_case_to_camel_case
 from spdx.writer.rdf.writer_utils import add_namespace_to_spdx_id
 from spdx.rdfschema.namespace import SPDX_NAMESPACE
