+++ src/spdx/writer/rdf/relationship_writer.py
@@ -15,7 +15,7 @@ from rdflib import Graph, BNode, RDF, URIRef
 from spdx.model.relationship import Relationship
 from spdx.model.spdx_no_assertion import SpdxNoAssertion
 from spdx.model.spdx_none import SpdxNone
-from spdx.writer.casing_tools import snake_case_to_camel_case
+from spdx.casing_tools import snake_case_to_camel_case
 from spdx.writer.rdf.writer_utils import add_namespace_to_spdx_id
 from spdx.rdfschema.namespace import SPDX_NAMESPACE
